@@ -11,6 +11,33 @@
 |
 */
 
-Route::prefix('authorization')->group(function() {
-    Route::get('/', 'AuthorizationController@index');
+use Modules\Authorization\Http\Controllers\AuthorizationController;
+
+Route::prefix('auth')->name('auth.')->group(function () {
+//    Route::post('register', [ AuthorizationController::class, 'register' ])->name('register');
+
+    Route::post('login', [ AuthorizationController::class, 'login' ])->name('login');
+//    Route::post('logout', [ AuthorizationController::class, 'logout' ])->name('logout');
+//
+//    Route::post('change-credentials', [AuthorizationController::class, 'changeCredentials'])->name('change_credentials');
+//    Route::post('change-password', [AuthorizationController::class, 'changePassword'])->name('change_password');
+//
+//    Route::post('reset-password', [AuthorizationController::class, 'resetPassword'])->name('reset_password');
+//    Route::post('resend-activation', [AuthorizationController::class, 'reSendActivationEmail'])->middleware('throttle:1,60')->name('resend_activation');
+//
+//    Route::get('/verify/{token}', [ AuthorizationController::class, 'verify'])->name('verify');
+//    Route::any('/reset/{token}', [ AuthorizationController::class, 'resetPasswordView'])->name('reset_password_view');
+//
+//    Route::prefix('providers')->name('providers.')->group(static function () {
+//        Route::post('/', [ AuthorizationController::class, 'providers']);
+//
+//        Route::get('/redirect/{provider}', [AuthorizationController::class, 'redirect'])->name('providers.redirect');
+//        Route::get('/callback/{provider}', [AuthorizationController::class, 'callback'])->name('providers.callback');
+//
+//        Route::post('/verify/{provider}', [AuthorizationController::class, 'verifyProvider'])->name('providers.verify');
+//    });
+
+    Route::middleware('auth:sanctum')->group(function () {
+        Route::post('/me', [ AuthorizationController::class, 'me'])->name('me');
+    });
 });
